requirements/base.txt
django-braces==1.3.1
django-reversion==1.8.0
django-sendfile
django-simple-captcha
Django==1.6.1
pillow==2.0
python-dateutil==1.5
South==0.8.4
